reportes_ms/requirements.txt
Flask
pymongo
flask-cors
pandas
reportlab
python-dotenv
